// src/api/zzirit/models/TypeFetchResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * zzirit API
 * 찌릿 API 명세서
 *
 * The version of the OpenAPI document: v0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypeFetchResponse
 */
export interface TypeFetchResponse {
    /**
     * 
     * @type {number}
     * @memberof TypeFetchResponse
     */
    typeId?: number;
    /**
     * 
     * @type {string}
     * @memberof TypeFetchResponse
     */
    name?: string;
}

/**
 * Check if a given object implements the TypeFetchResponse interface.
 */
export function instanceOfTypeFetchResponse(value: object): value is TypeFetchResponse {
    return true;
}

export function TypeFetchResponseFromJSON(json: any): TypeFetchResponse {
    return TypeFetchResponseFromJSONTyped(json, false);
}

export function TypeFetchResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypeFetchResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'typeId': json['typeId'] == null ? undefined : json['typeId'],
        'name': json['name'] == null ? undefined : json['name'],
    };
}

export function TypeFetchResponseToJSON(json: any): TypeFetchResponse {
    return TypeFetchResponseToJSONTyped(json, false);
}

export function TypeFetchResponseToJSONTyped(value?: TypeFetchResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'typeId': value['typeId'],
        'name': value['name'],
    };
}